pyproject.toml
[tool.poetry]
name = "aim-sso-example"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "~3.12.0"
fastapi = "~0.115.6"
uvicorn = "~0.32.1"
python-keycloak = "~4.7.3"
